+++ src/api/java/com/codetaylor/mc/artisanworktables/api/internal/config/IModuleToolsConfig.java
@@ -20,4 +20,6 @@ public interface IModuleToolsConfig {
 
   boolean enableToolTypeOreDictGroups();
 
+  boolean enableToolRepair();
+
 }